fix: make store in memory and file persistence

+++ wallet/interface.go
@@ -1,7 +1,7 @@
 package wallet
 
 type IWallet interface {
-	BondTransaction(string, string, string, float64) (string, error)
-	TransferTransaction(string, string, string, float64) (string, error)
+	BondTransaction(string, string, string, int64) (string, error)
+	TransferTransaction(string, string, string, int64) (string, error)
 	Address() string
 }
